bamboo-financial-api/requirements.txt
# Bamboo Financial API - Requirements
# Framework principal
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Base de données PostgreSQL
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Authentification et sécurité
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation et sérialisation
pydantic==2.5.0
email-validator==2.1.0

# Variables d'environnement
python-dotenv==1.0.0

# Utilitaires
httpx==0.25.2
requests==2.31.0

# Dates et temps
python-dateutil==2.8.2

# CORS et middleware
starlette==0.27.0

# Documentation
jinja2==3.1.2

# Gestion des sessions
itsdangerous==2.1.2

# Timezone management
pytz==2023.3

# Environment variables validation
pydantic-settings==2.1.0

# Async support pour PostgreSQL
asyncpg==0.29.0

# File handling
aiofiles==23.2.1

# JSON handling optimisé
ujson==5.8.0

# Production WSGI server
gunicorn==21.2.0